+++ ui/components/ItemDetail/CandidatureLba/WidgetCandidatureLba.js
@@ -3,9 +3,9 @@ import { useFormik } from "formik"
 import React, { useEffect, useState } from "react"
 import { getItemId } from "../../../utils/getItemId"
 import { string_wrapper as with_str } from "../../../utils/wrapper_utils"
-import CandidatureSpontaneeFailed from "./CandidatureSpontaneeFailed"
-import CandidatureSpontaneeNominalBodyFooter from "./CandidatureSpontaneeNominalBodyFooter"
-import CandidatureSpontaneeWorked from "./CandidatureSpontaneeWorked"
+import CandidatureLbaFailed from "./CandidatureLbaFailed"
+import CandidatureLbaModalBody from "./CandidatureLbaModalBody"
+import CandidatureLbaWorked from "./CandidatureLbaWorked"
 import { getInitialSchemaValues, getValidationSchema } from "./services/getSchema"
 import hasAlreadySubmittedCandidature from "./services/hasAlreadySubmittedCandidature"
 import submitCandidature from "./services/submitCandidature"
@@ -69,19 +69,12 @@ const WidgetCandidatureLba = (props) => {
       ) : (
         <form onSubmit={formik.handleSubmit}>
           {with_str(sendingState).amongst(["not_sent", "currently_sending"]) && (
-            <CandidatureSpontaneeNominalBodyFooter
-              formik={formik}
-              sendingState={sendingState}
-              company={props?.item?.company?.name}
-              item={props?.item}
-              kind={kind}
-              fromWidget={true}
-            />
+            <CandidatureLbaModalBody formik={formik} sendingState={sendingState} company={props?.item?.company?.name} item={props?.item} kind={kind} fromWidget={true} />
           )}
 
-          {with_str(sendingState).amongst(["ok_sent"]) && <CandidatureSpontaneeWorked kind={kind} email={formik.values.email} company={props?.item?.company?.name} />}
+          {with_str(sendingState).amongst(["ok_sent"]) && <CandidatureLbaWorked kind={kind} email={formik.values.email} company={props?.item?.company?.name} />}
 
-          {!with_str(sendingState).amongst(["not_sent", "ok_sent", "currently_sending"]) && <CandidatureSpontaneeFailed sendingState={sendingState} />}
+          {!with_str(sendingState).amongst(["not_sent", "ok_sent", "currently_sending"]) && <CandidatureLbaFailed sendingState={sendingState} />}
         </form>
       )}
     </Box>
