+++ ui/pages/postuler.js
@@ -1,7 +1,7 @@
 import React from "react"
 import Head from "next/head"
 import { NextSeo } from "next-seo"
-import WidgetPostuler from "../components/ItemDetail/CandidatureSpontanee/WidgetPostuler"
+import WidgetPostuler from "../components/ItemDetail/CandidatureLba/WidgetPostuler"
 
 const Postuler = () => {
   return (
